+++ src/helpers/settings.ts
@@ -1,3 +1,4 @@
 export default {
     BASE_URL: 'https://api.chift.eu',
+    DEFAULT_PAGE_SIZE: 100,
 };