+++ src/modules/accounting.ts
@@ -95,6 +95,7 @@ const accountingFactory = {
             method: 'get',
             url: '/consumers/{consumer_id}/accounting/analytic-plans',
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     getClients(
@@ -106,6 +107,7 @@ const accountingFactory = {
             method: 'get',
             url: '/consumers/{consumer_id}/accounting/clients',
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     createClient(
@@ -119,6 +121,7 @@ const accountingFactory = {
             url: '/consumers/{consumer_id}/accounting/clients',
             body: client,
             clientRequestId: options?.clientRequestId,
+            datalayer: options?.datalayer,
         };
     },
     getClient(
@@ -131,6 +134,7 @@ const accountingFactory = {
             method: 'get',
             url: `/consumers/{consumer_id}/accounting/clients/${clientId}`,
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     updateClient(
@@ -145,6 +149,7 @@ const accountingFactory = {
             url: `/consumers/{consumer_id}/accounting/clients/${clientId}`,
             body: client,
             clientRequestId: options?.clientRequestId,
+            datalayer: options?.datalayer,
         };
     },
     getSuppliers(
@@ -156,6 +161,7 @@ const accountingFactory = {
             method: 'get',
             url: '/consumers/{consumer_id}/accounting/suppliers',
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     createSupplier(
@@ -169,6 +175,7 @@ const accountingFactory = {
             url: '/consumers/{consumer_id}/accounting/suppliers',
             body: supplier,
             clientRequestId: options?.clientRequestId,
+            datalayer: options?.datalayer,
         };
     },
     getSupplier(
@@ -181,6 +188,7 @@ const accountingFactory = {
             method: 'get',
             url: `/consumers/{consumer_id}/accounting/suppliers/${supplierId}`,
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     updateSupplier(
@@ -195,6 +203,7 @@ const accountingFactory = {
             url: `/consumers/{consumer_id}/accounting/suppliers/${supplierId}`,
             body: supplier,
             clientRequestId: options?.clientRequestId,
+            datalayer: options?.datalayer,
         };
     },
     createInvoice(
@@ -208,6 +217,7 @@ const accountingFactory = {
             url: '/consumers/{consumer_id}/accounting/invoices',
             body: invoice,
             clientRequestId: options?.clientRequestId,
+            datalayer: options?.datalayer,
         };
     },
     createInvoiceWithMultiplePlans(
@@ -221,6 +231,7 @@ const accountingFactory = {
             url: '/consumers/{consumer_id}/accounting/invoices/multi-analytic-plans',
             body: invoice,
             clientRequestId: options?.clientRequestId,
+            datalayer: options?.datalayer,
         };
     },
     getInvoicesByType(
@@ -233,6 +244,7 @@ const accountingFactory = {
             method: 'get',
             url: `/consumers/{consumer_id}/accounting/invoices/type/${invoice_type}`,
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     getInvoice(
@@ -245,6 +257,7 @@ const accountingFactory = {
             method: 'get',
             url: `/consumers/{consumer_id}/accounting/invoices/${invoiceId}`,
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     getInvoiceWithMultiplePlans(
@@ -257,6 +270,7 @@ const accountingFactory = {
             method: 'get',
             url: `/consumers/{consumer_id}/accounting/invoices/multi-analytic-plans/${invoiceId}`,
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     getInvoicesByTypeWithMultiplePlans(
@@ -269,6 +283,7 @@ const accountingFactory = {
             method: 'get',
             url: `/consumers/{consumer_id}/accounting/invoices/multi-analytic-plans/type/${invoice_type}`,
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     createAnalyticAccount(
@@ -282,6 +297,7 @@ const accountingFactory = {
             url: '/consumers/{consumer_id}/accounting/analytic-accounts',
             body: analyticAccount,
             clientRequestId: options?.clientRequestId,
+            datalayer: options?.datalayer,
         };
     },
     getAnalyticAccounts(
@@ -293,6 +309,7 @@ const accountingFactory = {
             method: 'get',
             url: '/consumers/{consumer_id}/accounting/analytic-accounts',
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     createAnalyticAccountWithMultiplePlans(
@@ -307,6 +324,7 @@ const accountingFactory = {
             url: `/consumers/{consumer_id}/accounting/analytic-accounts/multi-analytic-plans/${analytic_plan}`,
             body: analyticAccount,
             clientRequestId: options?.clientRequestId,
+            datalayer: options?.datalayer,
         };
     },
     getAnalyticAccount(
@@ -319,6 +337,7 @@ const accountingFactory = {
             method: 'get',
             url: `/consumers/{consumer_id}/accounting/analytic-accounts/${analytic_account_id}`,
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     updateAnalyticAccount(
@@ -333,6 +352,7 @@ const accountingFactory = {
             url: `/consumers/{consumer_id}/accounting/analytic-accounts/${analytic_account_id}`,
             body: analyticAccount,
             clientRequestId: options?.clientRequestId,
+            datalayer: options?.datalayer,
         };
     },
     getAnalyticAccountWithMultiplePlans(
@@ -346,6 +366,7 @@ const accountingFactory = {
             method: 'get',
             url: `/consumers/{consumer_id}/accounting/analytic-accounts/${analytic_account_id}/multi-analytic-plans/${analytic_plan}`,
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     updateAnalyticAccountWithMultiplePlans(
@@ -370,6 +391,7 @@ const accountingFactory = {
             method: 'get',
             url: '/consumers/{consumer_id}/accounting/analytic-accounts/multi-analytic-plans',
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     getJournalEntries(
@@ -381,6 +403,7 @@ const accountingFactory = {
             method: 'get',
             url: `/consumers/{consumer_id}/accounting/journal/entries`,
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     getJournalEntriesWithMultiplePlans(
@@ -392,6 +415,7 @@ const accountingFactory = {
             method: 'get',
             url: `/consumers/{consumer_id}/accounting/journal/entries/multi-analytic-plans`,
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     getPaymentsByInvoiceId(
@@ -404,6 +428,7 @@ const accountingFactory = {
             method: 'get',
             url: `/consumers/{consumer_id}/accounting/invoices/id/${invoice_id}/payments`,
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     getJournals(
@@ -415,6 +440,7 @@ const accountingFactory = {
             method: 'get',
             url: `/consumers/{consumer_id}/accounting/journals`,
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     createJournal(
@@ -428,6 +454,7 @@ const accountingFactory = {
             url: '/consumers/{consumer_id}/accounting/journal',
             body: journal,
             clientRequestId: options?.clientRequestId,
+            datalayer: options?.datalayer,
         };
     },
     getVatCodes(
@@ -439,6 +466,7 @@ const accountingFactory = {
             method: 'get',
             url: `/consumers/{consumer_id}/accounting/vat-codes`,
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     getMiscOperations(
@@ -450,6 +478,7 @@ const accountingFactory = {
             method: 'get',
             url: '/consumers/{consumer_id}/accounting/miscellaneous-operation',
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     createMiscOperation(
@@ -463,6 +492,7 @@ const accountingFactory = {
             url: '/consumers/{consumer_id}/accounting/miscellaneous-operation',
             body: operation,
             clientRequestId: options?.clientRequestId,
+            datalayer: options?.datalayer,
         };
     },
     getMiscOperation(
@@ -475,6 +505,7 @@ const accountingFactory = {
             method: 'get',
             url: `/consumers/{consumer_id}/accounting/miscellaneous-operation/${operation_id}`,
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     attachPDF(
@@ -498,6 +529,7 @@ const accountingFactory = {
             method: 'get',
             url: `/consumers/{consumer_id}/accounting/attachments`,
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     getChartOfAccounts(
@@ -509,6 +541,7 @@ const accountingFactory = {
             method: 'get',
             url: '/consumers/{consumer_id}/accounting/chart-of-accounts',
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     getBalanceOfAccounts(
@@ -531,6 +564,7 @@ const accountingFactory = {
             method: 'get',
             url: '/consumers/{consumer_id}/accounting/employees',
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     getOutstandings(
@@ -542,6 +576,7 @@ const accountingFactory = {
             method: 'get',
             url: '/consumers/{consumer_id}/accounting/outstandings',
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     createFinancialEntry(
@@ -555,6 +590,7 @@ const accountingFactory = {
             url: '/consumers/{consumer_id}/accounting/financial-entries',
             body: financial_entry,
             clientRequestId: options?.clientRequestId,
+            datalayer: options?.datalayer,
         };
     },
     createJournalEntry(
@@ -568,6 +604,7 @@ const accountingFactory = {
             url: '/consumers/{consumer_id}/accounting/journal-entries',
             body: journal_entry,
             clientRequestId: options?.clientRequestId,
+            datalayer: options?.datalayer,
         };
     },
     matchEntries(
@@ -603,6 +640,7 @@ const accountingFactory = {
             url: '/consumers/{consumer_id}/accounting/expenses',
             body: expense,
             clientRequestId: options?.clientRequestId,
+            datalayer: options?.datalayer,
         };
     },
     getFolders(options?: RawDataOption): RequestData<components['schemas']['FolderItem'][]> {
@@ -610,6 +648,7 @@ const accountingFactory = {
             method: 'get',
             url: '/consumers/{consumer_id}/accounting/folders',
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     getBookyears(
@@ -621,6 +660,7 @@ const accountingFactory = {
             method: 'get',
             url: '/consumers/{consumer_id}/accounting/bookyears',
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     createLedgerAccount(
@@ -634,6 +674,7 @@ const accountingFactory = {
             url: '/consumers/{consumer_id}/accounting/accounts',
             body: account,
             clientRequestId: options?.clientRequestId,
+            datalayer: options?.datalayer,
         };
     },
     getBankAccounts(
@@ -645,6 +686,7 @@ const accountingFactory = {
             method: 'get',
             url: '/consumers/{consumer_id}/accounting/bank-accounts',
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     createBankAccount(
@@ -658,6 +700,7 @@ const accountingFactory = {
             url: '/consumers/{consumer_id}/accounting/bank-accounts',
             body: bankAccount,
             clientRequestId: options?.clientRequestId,
+            datalayer: options?.datalayer,
         };
     },
     createBankTransactions(
@@ -671,6 +714,7 @@ const accountingFactory = {
             url: '/consumers/{consumer_id}/accounting/bank-transactions',
             body: bankStatement,
             clientRequestId: options?.clientRequestId,
+            datalayer: options?.datalayer,
         };
     },
     getJournalEntry(
@@ -683,6 +727,7 @@ const accountingFactory = {
             method: 'get',
             url: `/consumers/{consumer_id}/accounting/journal/entries/${journalEntryId}`,
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
 
@@ -695,6 +740,7 @@ const accountingFactory = {
             method: 'get',
             url: `/consumers/{consumer_id}/accounting/payment-methods`,
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     getPaymentTerms(
@@ -706,6 +752,7 @@ const accountingFactory = {
             method: 'get',
             url: `/consumers/{consumer_id}/accounting/payment-terms`,
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     getSchemes(
@@ -717,6 +764,7 @@ const accountingFactory = {
             method: 'get',
             url: `/consumers/{consumer_id}/accounting/schemes`,
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
 
@@ -740,6 +788,7 @@ const accountingFactory = {
             method: 'get',
             url: `/consumers/{consumer_id}/accounting/export-fec`,
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
 };