+++ src/modules/banking.ts
@@ -31,6 +31,7 @@ const bankingFactory = {
             method: 'get',
             url: '/consumers/{consumer_id}/banking/financial-institutions',
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     getAccounts(
@@ -42,6 +43,7 @@ const bankingFactory = {
             method: 'get',
             url: '/consumers/{consumer_id}/banking/accounts',
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     getAccountTransactions(
@@ -53,6 +55,7 @@ const bankingFactory = {
             method: 'get',
             url: `/consumers/{consumer_id}/banking/transactions`,
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     getAccountCounterparts(
@@ -64,6 +67,7 @@ const bankingFactory = {
             method: 'get',
             url: '/consumers/{consumer_id}/banking/counterparts',
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
     getAttachments(
@@ -75,6 +79,7 @@ const bankingFactory = {
             method: 'get',
             url: '/consumers/{consumer_id}/banking/attachments',
             rawData: options?.rawData,
+            datalayer: options?.datalayer,
         };
     },
 };